Restructured lists functional tests into separate files
