+++ functional_tests/lists/test_lists_layout_and_styling.py
@@ -1,0 +1,17 @@
+from .lists_base import FunctionalTest
+
+
+class LayoutAndStylingTest(FunctionalTest):
+    def test_layout_and_styling(self):
+        # Edith goes to the home page
+        lists_server_url = '%s%s' % (self.server_url, '/lists')
+        self.browser.get(lists_server_url)
+        self.browser.set_window_size(800, 600)
+
+        # She notices the input box is nicely centered
+        inputbox = self.browser.find_element_by_id('id_new_item')
+        self.assertAlmostEqual(
+            inputbox.location['x'] + inputbox.size['width'] / 2,
+            400,
+            delta=5
+        )
