+++ functional_tests/lists/test_lists_simple_list_creation.py
@@ -1,47 +1,13 @@
-__author__ = 'cbinder'
-
+from .lists_base import FunctionalTest
 from selenium import webdriver
 from selenium.webdriver.common.keys import Keys
-from django.test import LiveServerTestCase
+#from django.test import LiveServerTestCase
 from django.contrib.staticfiles.testing import StaticLiveServerTestCase
 import time, sys
 
 
-class NewVisitorTest(StaticLiveServerTestCase):
-
-    @classmethod
-    def setUpClass(cls):
-        for arg in sys.argv:
-            if 'liveserver' in arg:
-                cls.server_url = 'http://' + arg.split('=')[1]
-                return
-        super().setUpClass()
-        cls.server_url = cls.live_server_url
-
-    @classmethod
-    def tearDownClass(cls):
-        if cls.server_url == cls.live_server_url:
-            super().tearDownClass()
-
-    def setUp(self):  #2
-        #chrome_option = webdriver.ChromeOptions()
-        #chrome_option.add_argument('--proxy-server=us-auto.proxy.lexmark.com:80' )
-        self.browser = webdriver.Chrome()
-        self.browser.implicitly_wait(3)
-        #self.browser.service()
-        #self.browser.
-
-    def tearDown(self):  #3
-        self.browser.close()
-        #self.browser.quit()
-
-    # helper function
-    def check_for_row_in_list_table(self, row_text):
-        table = self.browser.find_element_by_id('id_list_table')
-        rows = table.find_elements_by_tag_name('tr')
-        self.assertIn(row_text, [row.text for row in rows])
-
-    def test_can_start_a_list_and_retrieve_it_later(self):  #4
+class NewVisitorTest(FunctionalTest):
+    def test_can_start_a_list_and_retrieve_it_later(self):
         # Edith has heard about a cool new online to-do app. She goes
         # to check out its homepage
         #self.browser.get('http://localhost:8000/lists')
@@ -122,17 +88,3 @@ class NewVisitorTest(StaticLiveServerTestCase):
         time.sleep(2)
 
         # Satisfied, they both go back to sleep
-
-    def test_layout_and_styling(self):
-        # Edith goes to the home page
-        lists_server_url = '%s%s' % (self.server_url, '/lists')
-        self.browser.get(lists_server_url)
-        self.browser.set_window_size(800, 600)
-
-        # She notices the input box is nicely centered
-        inputbox = self.browser.find_element_by_id('id_new_item')
-        self.assertAlmostEqual(
-            inputbox.location['x'] + inputbox.size['width'] / 2,
-            400,
-            delta=5
-        )
